requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
chex==0.1.7
comm==0.2.0
contourpy==1.0.7
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
dm-haiku==0.0.11
dm-tree==0.1.8
etils==1.5.2
executing==1.2.0
flax==0.7.5
fonttools==4.39.4
fsspec==2023.10.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.13.2
jax==0.4.20
jaxlib==0.4.20+cuda12.cudnn89
jaxtyping==0.2.19
jedi==0.18.2
jmp==0.0.4
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
numpy==1.24.3
nvidia-cublas-cu12==12.3.2.9
nvidia-cuda-cupti-cu12==12.3.52
nvidia-cuda-nvcc-cu12==12.3.52
nvidia-cuda-nvrtc-cu12==12.3.52
nvidia-cuda-runtime-cu12==12.3.52
nvidia-cudnn-cu12==8.9.6.50
nvidia-cufft-cu12==11.0.11.19
nvidia-cusolver-cu12==11.5.3.52
nvidia-cusparse-cu12==12.1.3.153
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.52
opt-einsum==3.3.0
optax==0.1.5
orbax-checkpoint==0.4.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==4.0.0
prompt-toolkit==3.0.38
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
rich==13.7.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorstore==0.1.50
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typeguard==4.0.0
typing_extensions==4.5.0
wcwidth==0.2.6
zipp==3.17.0
